fix(ec2): handle attached state during EBS volume detachment wait
