+++ internal/service/ec2/ebs_volume_attachment_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/hashicorp/terraform-provider-aws/internal/acctest"
+	"github.com/hashicorp/terraform-provider-aws/internal/enum"
 	"github.com/hashicorp/terraform-provider-aws/internal/retry"
 	tfec2 "github.com/hashicorp/terraform-provider-aws/internal/service/ec2"
 	"github.com/hashicorp/terraform-provider-aws/names"
@@ -409,5 +410,69 @@ func testAccVolumeAttachmentImportStateIDFunc(resourceName string) resource.Impo
 			return "", fmt.Errorf("Not found: %s", resourceName)
 		}
 		return fmt.Sprintf("%s:%s:%s", rs.Primary.Attributes[names.AttrDeviceName], rs.Primary.Attributes["volume_id"], rs.Primary.Attributes[names.AttrInstanceID]), nil
+	}
+}
+
+// sequenceRefreshFunc returns a StateRefreshFunc that walks through states one
+// by one on each call. The final call returns an empty state (not-found).
+func sequenceRefreshFunc(states []awstypes.VolumeAttachmentState) retry.StateRefreshFunc {
+	i := 0
+	return func(_ context.Context) (any, string, error) {
+		if i >= len(states) {
+			return nil, "", nil
+		}
+		s := states[i]
+		i++
+		return &awstypes.VolumeAttachment{State: s}, string(s), nil
+	}
+}
+
+// TestVolumeAttachmentWaiterDeleted_attachedTransitionSucceeds verifies that the
+// waiter handles the attached → detaching → gone sequence without error.
+// Regression test for https://github.com/hashicorp/terraform-provider-aws/issues/47314:
+// AWS does not always immediately transition to detaching after DetachVolume is called.
+func TestVolumeAttachmentWaiterDeleted_attachedTransitionSucceeds(t *testing.T) {
+	t.Parallel()
+
+	conf := &retry.StateChangeConf{
+		Pending: enum.Slice(
+			awstypes.VolumeAttachmentStateAttached,
+			awstypes.VolumeAttachmentStateDetaching,
+		),
+		Target: []string{},
+		Refresh: sequenceRefreshFunc([]awstypes.VolumeAttachmentState{
+			awstypes.VolumeAttachmentStateAttached,
+			awstypes.VolumeAttachmentStateDetaching,
+		}),
+		Timeout:    5 * time.Second,
+		Delay:      0,
+		MinTimeout: 0,
+	}
+
+	_, err := conf.WaitForStateContext(context.Background())
+	if err != nil {
+		t.Fatalf("expected no error, got: %s", err)
+	}
+}
+
+// TestVolumeAttachmentWaiterDeleted_attachedOnlyPending_fails shows the pre-fix
+// behaviour: with only detaching in Pending, seeing attached immediately fails.
+func TestVolumeAttachmentWaiterDeleted_attachedOnlyPending_fails(t *testing.T) {
+	t.Parallel()
+
+	conf := &retry.StateChangeConf{
+		Pending: enum.Slice(awstypes.VolumeAttachmentStateDetaching),
+		Target:  []string{},
+		Refresh: sequenceRefreshFunc([]awstypes.VolumeAttachmentState{
+			awstypes.VolumeAttachmentStateAttached,
+		}),
+		Timeout:    5 * time.Second,
+		Delay:      0,
+		MinTimeout: 0,
+	}
+
+	_, err := conf.WaitForStateContext(context.Background())
+	if err == nil {
+		t.Fatal("expected an error for unexpected 'attached' state, got none")
 	}
 }
